refactor: extract shared modal styles into reusable component
ModalActions, CancelButton, DoneButton, TextFieldLabel, ModalBox,
ModalTitle, ModalDescription, ModalContainer, and ModalContent were
duplicated across 13 modal styles files. This extracts the common
definitions into src/components/Modal/styles.ts and updates each
consumer to re-export from shared, keeping only local overrides.
