+++ src/pages/Participants/AddNewDeploymentModal/styles.ts
@@ -1,6 +1,4 @@
 import {
-  Box,
-  Button,
   Divider,
   Select,
   TableCell,
@@ -9,20 +7,21 @@ import {
   Typography,
 } from '@mui/material';
 import { styled } from '@Utils/theme';
+import { ModalBox as BaseModalBox } from '@Components/Modal/styles';
 
-export const ModalBox = styled(Box)(({ theme }) => ({
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: '70%',
-  backgroundColor: theme.palette.common.white,
-  borderRadius: 16,
+export {
+  CancelButton,
+  DoneButton,
+  ModalActions,
+  TextFieldLabel,
+} from '@Components/Modal/styles';
+
+export const ModalBox = styled(BaseModalBox)({
   padding: 24,
   maxWidth: 800,
   display: 'flex',
   flexDirection: 'column',
-}));
+});
 
 export const ModalTitle = styled(Typography)(({ theme }) => ({
   color: theme.palette.primary.main,
@@ -60,38 +59,6 @@ export const StyledTableContainer = styled(TableContainer)({
 export const ModalContent = styled('div')({
   flexGrow: 2,
 });
-
-export const TextFieldLabel = styled('label')(({ theme }) => ({
-  marginTop: 12,
-  color: theme.palette.secondary.dark,
-  fontSize: theme.typography.h4.fontSize,
-  fontWeight: theme.typography.h4.fontWeight,
-  lineHeight: theme.typography.h4.lineHeight,
-}));
-
-export const ModalActions = styled('div')({
-  display: 'flex',
-  flexDirection: 'row',
-  gap: 8,
-  marginTop: 24,
-  width: '100%',
-  justifyContent: 'flex-end',
-});
-
-export const CancelButton = styled(Button)(({ theme }) => ({
-  color: theme.palette.primary.main,
-  textTransform: 'none',
-  padding: '8px 24px',
-  borderRadius: 16,
-}));
-
-export const DoneButton = styled(Button)(({ theme }) => ({
-  backgroundColor: theme.palette.primary.main,
-  color: theme.palette.common.white,
-  textTransform: 'none',
-  padding: '8px 24px',
-  borderRadius: 16,
-}));
 
 export const HeaderTableCell = styled(TableCell)(({ theme }) => ({
   borderBottomColor: theme.palette.grey[500],
