+++ src/pages/StudySettings/AddResearcherModal/styles.ts
@@ -1,17 +1,18 @@
-import { Box, Button, Typography } from '@mui/material';
+import { Typography } from '@mui/material';
 import { styled } from '@Utils/theme';
+import { ModalBox as BaseModalBox } from '@Components/Modal/styles';
 
-export const ModalBox = styled(Box)(({ theme }) => ({
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: '70%',
-  backgroundColor: theme.palette.common.white,
-  borderRadius: 16,
-  padding: 24,
+export {
+  CancelButton,
+  DoneButton,
+  ModalActions,
+  TextFieldLabel,
+} from '@Components/Modal/styles';
+
+export const ModalBox = styled(BaseModalBox)({
   maxWidth: 550,
-}));
+  padding: 24,
+});
 
 export const ModalTitle = styled(Typography)(({ theme }) => ({
   color: theme.palette.primary.main,
@@ -32,35 +33,3 @@ export const ModalContent = styled('div')({
   display: 'flex',
   flexDirection: 'column',
 });
-
-export const TextFieldLabel = styled('label')(({ theme }) => ({
-  marginTop: 12,
-  color: theme.palette.secondary.dark,
-  fontSize: theme.typography.h4.fontSize,
-  fontWeight: theme.typography.h4.fontWeight,
-  lineHeight: theme.typography.h4.lineHeight,
-}));
-
-export const ModalActions = styled('div')({
-  display: 'flex',
-  flexDirection: 'row',
-  gap: 8,
-  marginTop: 24,
-  width: '100%',
-  justifyContent: 'flex-end',
-});
-
-export const CancelButton = styled(Button)(({ theme }) => ({
-  color: theme.palette.primary.main,
-  textTransform: 'none',
-  padding: '8px 24px',
-  borderRadius: 16,
-}));
-
-export const DoneButton = styled(Button)(({ theme }) => ({
-  backgroundColor: theme.palette.primary.main,
-  color: theme.palette.common.white,
-  textTransform: 'none',
-  padding: '8px 24px',
-  borderRadius: 16,
-}));
